backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0
python-dotenv>=1.0
firecrawl>=1.0
google-genai>=1.0
pydantic>=2.0
